Status record check

+++ geoportaal-beheer/app/controllers/Index.java
@@ -22,6 +22,7 @@ import com.querydsl.sql.SQLQuery;
 import actions.DefaultAuthenticator;
 import models.Delete;
 import models.Search;
+import models.Status;
 import models.Supplier;
 import play.Routes;
 import play.data.DynamicForm;
@@ -70,20 +71,35 @@ public class Index extends Controller {
 		});
     }
 	
-	public Result changeStatus(String metadataUuid, String statusStr) {
+	public Result changeStatus() {
+		Form<models.Status> statusForm = Form.form(models.Status.class);
+		models.Status s = statusForm.bindFromRequest().get();
+		List<String> changeRecords = s.getRecordsChange();
+		String statusName = s.getStatus();
+		
 		return q.withTransaction(tx -> {
-			Integer statusKey = tx.select(status.id)
-				.from(status)
-				.where(status.name.eq(statusStr))
-				.fetchFirst();
+			if(changeRecords != null && statusName != null) {
+				Integer statusKey = tx.select(status.id)
+					.from(status)
+					.where(status.name.eq(statusName))
+					.fetchOne();
 			
-			tx.update(metadata)
-	    		.where(metadata.uuid.eq(metadataUuid))
-	    		.set(metadata.status, statusKey)
-	    		.execute();
+				if(statusKey != null) {
+					Long count = tx.update(metadata)
+			    		.where(metadata.uuid.in(changeRecords))
+			    		.set(metadata.status, statusKey)
+			    		.execute();
+					
+					Integer finalCount = count.intValue();
+					if(!finalCount.equals(changeRecords.size())) {
+						throw new Exception("Changing status: different amount of affected rows than expected");
+					}
+				}
+			}
 	    	
 	    	return redirect(controllers.routes.Index.index());
 		});
+		
 	}
 	
 	public Result changeSupplier() {
@@ -93,19 +109,23 @@ public class Index extends Controller {
 		String supplierName = s.getSupplier();
 		
 		return q.withTransaction(tx -> {
-			Integer supplierKey = tx.select(supplier.id)
-				.from(supplier)
-				.where(supplier.name.eq(supplierName))
-				.fetchOne();
-			
-			Long count = tx.update(metadata)
-	    		.where(metadata.uuid.in(changeRecords))
-	    		.set(metadata.supplier, supplierKey)
-	    		.execute();
-	    	
-			Integer finalCount = count.intValue();
-			if(!finalCount.equals(changeRecords.size())) {
-				throw new Exception("Changing supplier: different amount of affected rows than expected");
+			if(changeRecords != null && supplierName != null) {
+				Integer supplierKey = tx.select(supplier.id)
+					.from(supplier)
+					.where(supplier.name.eq(supplierName))
+					.fetchOne();
+				
+				if(supplierKey != null) {
+					Long count = tx.update(metadata)
+			    		.where(metadata.uuid.in(changeRecords))
+			    		.set(metadata.supplier, supplierKey)
+			    		.execute();
+			    	
+					Integer finalCount = count.intValue();
+					if(!finalCount.equals(changeRecords.size())) {
+						throw new Exception("Changing supplier: different amount of affected rows than expected");
+					}
+				}
 			}
 			
 	    	return redirect(controllers.routes.Index.index());
@@ -119,24 +139,26 @@ public class Index extends Controller {
 		String permDel = d.getPermDel();
 		
 		return q.withTransaction(tx -> {
-			if(permDel != null) {
-				Long count = tx.delete(metadata)
-					.where(metadata.uuid.in(deleteRecords))
-					.execute();
-				
-				Integer finalCount = count.intValue();
-				if(!finalCount.equals(deleteRecords.size())) {
-					throw new Exception("Deleting records: different amount of affected rows than expected");
-				}
-			} else {
-				Long count = tx.update(metadata)
-					.where(metadata.uuid.in(deleteRecords))
-					.set(metadata.status, 5)
-					.execute();
-				
-				Integer finalCount = count.intValue();
-				if(!finalCount.equals(deleteRecords.size())) {
-					throw new Exception("Change status to deleted: different amount of affected rows than expected");
+			if(deleteRecords != null) {
+				if(permDel != null) {
+					Long count = tx.delete(metadata)
+						.where(metadata.uuid.in(deleteRecords))
+						.execute();
+					
+					Integer finalCount = count.intValue();
+					if(!finalCount.equals(deleteRecords.size())) {
+						throw new Exception("Deleting records: different amount of affected rows than expected");
+					}
+				} else {
+					Long count = tx.update(metadata)
+						.where(metadata.uuid.in(deleteRecords))
+						.set(metadata.status, 5)
+						.execute();
+					
+					Integer finalCount = count.intValue();
+					if(!finalCount.equals(deleteRecords.size())) {
+						throw new Exception("Change status to deleted: different amount of affected rows than expected");
+					}
 				}
 			}
 			
@@ -281,7 +303,7 @@ public class Index extends Controller {
             controllers.routes.javascript.Index.deleteMetadata(),
 			controllers.routes.javascript.Index.changeStatus(),
 			controllers.routes.javascript.Index.changeSupplier(),
-			controllers.routes.javascript.MetadataDC.validateForm(),
+			controllers.routes.javascript.Metadata.validateForm(),
 			controllers.routes.javascript.Index.validateForm()
         )).as ("text/javascript");
     }
